spline #605 Refactoring: move `extractChildren(Node)` method from `LineageHarvester` to `OperationNodeBuilderFactory`, and delegate `MergeIntoCommand` children extraction logic to the `MergeIntoNodeBuilder` respectively.
